images_converter/requirements.txt
Pillow==9.0


